+++ src/main/java/org/sitmun/domain/territory/type/TerritoryType.java
@@ -8,9 +8,12 @@ import java.util.Objects;
 import lombok.*;
 import org.sitmun.authorization.client.dto.ClientConfigurationViews;
 import org.sitmun.domain.PersistenceConstants;
+import org.sitmun.infrastructure.persistence.type.i18n.I18n;
+import org.sitmun.infrastructure.persistence.type.i18n.I18nListener;
 
 /** Type of territorial entities. */
 @Entity
+@EntityListeners(I18nListener.class)
 @Table(
     name = "STM_TER_TYP",
     uniqueConstraints = @UniqueConstraint(name = "STM_TET_NOM_UK", columnNames = "TET_NAME"))
@@ -37,6 +40,7 @@ public class TerritoryType {
   /** Name. */
   @Column(name = "TET_NAME", length = PersistenceConstants.IDENTIFIER)
   @NotBlank
+  @I18n
   @JsonView(ClientConfigurationViews.ApplicationTerritory.class)
   private String name;